Attempted to make the database save data from the form
I have saved my code but commented it out. Its going in the correct direction just cant seem to get it to grab the data from the form.I think its the way the form is set up. I have not tried javascript only tried to call the route to the form. if you want to save to site practice you need to change the APP.CONFIG at the top of the page from site.db to sitePractice.db. ENJOY